+++ server/src/course/course.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
+import { CourseController } from './course.controller';
+import { CourseService } from './course.service';
+
+@Module({
+  imports: [PrismaModule],
+  controllers: [CourseController],
+  providers: [CourseService],
+})
+export class CourseModule {}